Refactor to also use resumable HTTP input stream and User-Agent in download command
